MCP-1679: simplify test using camel mocks
